+++ crates/argmin-math/src/ndarray_m/mod.rs
@@ -16,9 +16,15 @@ mod eye;
     feature = "ndarray-linalg_0_12",
     feature = "ndarray-linalg_0_13",
     feature = "ndarray-linalg_0_16",
-    feature = "faer"
 ))]
 mod inv;
+#[cfg(all(not(
+    any(feature = "ndarray-linalg_0_12", 
+        feature = "ndarray-linalg_0_13", 
+        feature = "ndarray-linalg_0_16")), 
+    feature = "faer", 
+    ))]
+mod inv_faer;
 mod l1norm;
 mod l2norm;
 mod minmax;
@@ -39,9 +45,16 @@ pub use eye::*;
 #[cfg(any(
     feature = "ndarray-linalg_0_12",
     feature = "ndarray-linalg_0_13",
-    feature = "ndarray-linalg_0_16"
+    feature = "ndarray-linalg_0_16",
 ))]
 pub use inv::*;
+#[cfg(all(not(
+    any(feature = "ndarray-linalg_0_12", 
+        feature = "ndarray-linalg_0_13", 
+        feature = "ndarray-linalg_0_16")), 
+    feature = "faer", 
+    ))]
+pub use inv_faer::*;
 pub use l1norm::*;
 pub use l2norm::*;
 pub use minmax::*;
